Created 'Add To Slack' button
models.py: Created the Slack 'Team' model.
views.py: The Slack client ID is now passed to the Home view so that it can be used by the 'Add To Slack' button.
home.html: Added the 'Add To Slack' button.
admin.py: Added the Team model to the Admin site.
views.py: Added the 'oauth' method which is used for authorising the app on the Slack team
urls.py: Added the 'oauth' URL pattern.
callback.py: Corrected the syntax for getting values from the Django settings file.